SCC-236
- Removed getRefileRequestByBarcode
- Moved afMessage and sip2Response to RefileRequestModel
- Modified logic of setting success flag to successful only when Alert=N and Ok=1.
