+++ src/Model/RefileRequestModel.php
@@ -15,6 +15,18 @@ class RefileRequestModel extends Model
     public $itemBarcode;
 
     /**
+     * @SWG\Property(example="Item was put on holdshelf."")
+     * @var string
+     */
+    public $afMessage;
+
+    /**
+     * @SWG\Property(example="{"fixed":{}, "variable": {}}")
+     * @var string
+     */
+    public $sip2Response;
+
+    /**
      * @return string
      */
     public function getItemBarcode()
@@ -28,5 +40,38 @@ class RefileRequestModel extends Model
     public function setItemBarcode($itemBarcode)
     {
         $this->itemBarcode = $itemBarcode;
+    }
+
+    /**
+     * @return string
+     */
+    public function getAfMessage()
+    {
+        return $this->afMessage;
+    }
+
+    /**
+     * @param string $afMessage
+     */
+    public function setAfMessage($afMessage)
+    {
+        $this->afMessage = $afMessage;
+    }
+
+    /**
+     * @return string
+     */
+    public function getSip2Response()
+    {
+        return json_decode($this->sip2Response) ;
+    }
+
+    /**
+     * @param string $sip2Response
+     */
+    public function setSip2Response($sip2Response)
+    {
+
+        $this->sip2Response = $sip2Response;
     }
 }
